work related to generation of synthetic images for neural net training

+++ visnav/missions/didymos.py
@@ -41,8 +41,8 @@ class DidymosSystemModel(SystemModel):
         common_kwargs = common_kwargs_best
 
         narrow_cam = Camera(
-            2048 * res_mult,   # width in pixels
-            1944 * res_mult,   # height in pixels
+            int(2048 * res_mult),   # width in pixels
+            int(1944 * res_mult),   # height in pixels
             7.7,    # x fov in degrees  (could be 6 & 5.695, 5.15 & 4.89, 7.7 & 7.309)
             7.309,  # y fov in degrees
             f_stop=5,       # TODO: put better value here
@@ -51,8 +51,8 @@ class DidymosSystemModel(SystemModel):
             **common_kwargs
         )
         wide_cam = Camera(
-            2048 * res_mult,   # width in pixels
-            1944 * res_mult,   # height in pixels
+            int(2048 * res_mult),   # width in pixels
+            int(1944 * res_mult),   # height in pixels
             61.5,     # x fov in degrees
             58.38,  # y fov in degrees
             f_stop=5,   # TODO: put better value here
